Move diff listener to impl
